Synchronize threads
Once all threads are started, program start is set 100ms in the future.
All threads wait to this time before doing their things.
This makes threads do their actions with much tighter timing, ~3 ms for 200 philos.
This was done because at eval, you are given 10 ms of wiggle room for up to 200 philos